Update libxml2 windows project files
The libxml2 windows project files will now put the resultant libxml2
libraries in the Ema/Libs directory so you don't need to copy them to
that location manually.

In addition, this update fixes a misspelling of libxml2 for VS2012 Debug
Shared builds. Previously, the library name was missing the 'l' between
the 'm' and '2'.