ETA-2321 Fix Reactor 64-connection limit
Add a notifier utility that can handle many more connections.
On Linux, this uses poll() and is able to grow as needed. On Windows,
select() is still used because WSAPoll() does not report failed
connections (blocking the reactor from reconnecting them). However,
FD_SETSIZE was increased to 6400.

Fix upacProvPerf not destroying the ProviderSession object when
rsslReactorAccept fails, resulting in a memory leak.