ESDK-1300 Refactor ETA Java Reactor Watchlist item recovery/cleanup
Refactor code in WlItemHandler that is used to recover and cleanup
closed items, and fix instances of double-repooling of objects or members
of those objects being cleaned up in improper order:

- Collapse item recovery code

Collapse item recovery code so that handling of closed streams and
subsequent message fanout goes through a single code path (replaces
methods such as handleCloseRecover, handleClose, handleRedirected).

- Make closing/cleanup of item stream objects consistent

Add WlItemHandler.closeWlStream method so that WlStreams are cleaned up
consistently when item streams are closed. Remove WlStreams from their
respective watchlist tables and lists prior to sending callback messages
and repool them aftewards.

Similarly, as the watchlist fans out closed statuses to their requests,
remove the corresponding WlRequest from the watchlist table prior to the
callback, and repool it aftewards. Add unit test streamReopenTest for
testing this.

Add an assertion to catch double-repooling of WlRequests and fix a case
of this revealed by the loginClosedRecoverTest_SingleOpenOff test.

Make VaNode.inPool public and use it in an assertion to catch a WlStream
being closed or repooled after already being repooled.

- Fix order of WlView vs. WlRequest cleanup

Rename WlItemHandler.putWlRequestViewListBackToPool to
WlItemHandler.repoolWlRequest, and have the method repool other members
of the WlRequest, and then the WlRequest itself (previously a WlRequest
may be pooled before its associated View object).  Use this in place of instances of
code that were repooling the request object.

Add method WlRequest.returnToPool (overriding VaNode.returnToPool),
which sets WlRequest.state to RETURN_TO_POOL before pooling the request,
so that instances of code that repool the WlRequest don't have to do
this. This method also repools WlRequest.tableKey if present.

- Fix item group management

Add a unit test to test management of streams in item group merges.

Change WlStreams to point to the WlItemGroup itself instead of the
GroupId buffer, so that when the group ID changes (i.e. a group merge),
the streams will still be associated with the correct group (previously, they were
still pointing to the original GroupId buffer).

Add and use methods that add/remove WlStreams from their respective item
groups, so that they are properly removed from their previous group
when it changes.

- Fix SymbolList data stream removal & cleanup

Add a unit test symbolListDataSnapshotTest.  Fix a case where a
SymbolList request requesting data snapshots is not removed from
watchlist tables if the symbol list itself is a snapshot, causing it to
be recovered.

Fix watchlist not opening data streams for a SymbolList when when the
SymbolList response is a snapshot (the message payload was ignored in this
case).

Add unit test symbolListDataStreamUpdateSymbolListTest to test removal
of items from the table of open SymbolList data streams.  Ensure
WlRequest objects for those streams are consistently removed from the
table of provider-driven requests, so that an item will be reopened if
it appears in a SymbolList message after being closed.

Add method WlItemhandler.closeWlRequest that removes provider-driven
requests from the table, and use it when the request is closed by either
the consumer or the provider.

- Fix instance of WlRequest double-close

Add an assertion to check that a WlRequest was present in the watchlist
ID-to-request table when removing it, to ensure we are not
double-closing it.  Fix case where a nonstreaming refresh resulted in a
double-close of a WlRequest.

- Add unit test viewAggregateAndReconnect to test aggregated view
recovery on disconnection to ensure this works properly.

Fix Watchlist returning an error if a view is reissued without an
encodedDataBody (a RequestMsg with a View flag but no payload means the
view should be maintained).  This occurred because the watchlist view
handling set the original encodedDataBody without setting the
containerType (leaving it as NO_DATA).  Fixed by setting the
containerType to ELEMENT_LIST in that case.

- Add unit test openWindowReconnectTest to ensure ensure requests
waiting for an opening in an service's OpenWindow received proper status
and were recovered on disconnection.

Previously, the WlService object used in handleCloseRecover was already
pooled when its waitingRequestList was checked.  When services are
removed (e.g. due to a disconnect), ensure item status fanout is done
before each service is repooled, and that WlStreams are removed from
their respective WlServices so that they do not attempt to resubmit
requests that were waiting for an opening in an OpenWindow.

- Fix watchlistItemRecoveryDueToDirectoryStreamCloseTest

The test was previously not actually getting an item status that it
expected (the event it got was actually the DirectoryUpdate), but now does get
that status. Fixed the test to validate both the directory update and item status.

- Fix watchlist fanout of login close message

Add testing for closing the login stream from the consumer and provider
while an item stream is open (tests send an item update from the
provider after the close to confirm that the item is no longer in
the watchlist).

Fix the consumer's login CloseMsg being fanned out back to its item
requests (send a status instead).

- Fanout unrecognized stream states

Treat unrecognized StreamStates to be non-open and non-recoverable, and
fan them out (previously, messages with these states were silently
dropped). Change itemGroupUpdatedTest to account for this change (use
non-final StreamStates until the end, and make the test expect the
fanned out StatusMsg).

- Fix stream ID of StatusMsg with no State

In emptyStatusMsgTest, check the Stream ID of the StatusMsg.
Previously, StatusMsgs without a State were forwarded with
an incorrect stream ID.

- Add Gradle info for wlconsumer-MutiThreaded-001

The QA Tool wlconsumer-MutiThreaded-001 requires the XmlItemInfoList
object from the performance tools. Add info to the QATools on how to add
this dependency in Gradle (includes instructions in the QATools readme
and a build.gradle file).