ESDK-1365 ItemCallbackClient NullPointerException when rsslMsg state is null
The TunnelStream class of ETAJ has a member variable called _decSubMsg to decode sub-stream messages for both incoming from network and submitting messages by the application layer. That's why the _decSubMsg is decoded as Refresh message when receiving message from network then it is decoded again as Post message when the application submits a post message within the onRefreshMsg() callback method of EMA. This causes Java runtime to throw NullPointerException when EMA gets the state from Msg which has been changed from Refresh to Post message as the state no longer exists(null).

The solution for this issue to isolate the ETAJ Msg that is used for decoding buffer for receiving and submitting sides to avoid message overwritten within the defaultMsgCallback() call from the reactor.