ESDK-1461 Fix ETAJ Watchlist extra message with snapshot after streaming
If a snapshot is requested while a streaming request for the same item is
waiting for a refresh, the watchlist queued the item stream even though nothing
needs to be sent, resulting in a redundant request when the WlItemHandler is
dispatched.

Added Junit testing for this case.  Added similar testing for aggregation of
two streaming requests.