ESDK-1489 Make Msg.encodeComplete return error on NO_DATA with payload
If Msg.encodeComplete is called on a Msg whose containerType is NO_DATA, return
an error if the application encoded payload.

If Msg.encodeInit or Msg.encodeMsgKeyAttribComplete is called when the
containerType is NO_DATA, return SUCCESS instead of ENCODE_CONTAINER
(Msg.encodeExtendedHeaderComplete does this currently).

Add a Junit test that encodes a message with containerType NO_DATA with and
without payload, based on the noDataTest in ETAC.