ESDK-1496
merging changes for ESDK-1496 into develop. Merge was done by hand because automatic merge from
the branch was a mess.

fixed an issue in QA test tools. The timing of the reissues messages was supposed to be managed
by the provider but in some cases the reissue time reset message was received too late and an
additional reissue was sent for the previous reissue time. Fixed by resetting reissue time to zero
upon sending reissue message.