ESDK-1423 Compiler warnings, rsslCopyMsg crash
Fix warnings when compiling ESDK on Windows:

- Fix signed/unsigned mismatch warnings in rsslCopyMsg code that creates
  API-allocated copies. Add the buffer length to the subtracted
  pointers, so that the final result is unsigned.

- Fix signed/unsigned compiler warning in
  VAConsumer/rsslDirectoryHandler.c when checking the upper limit of the
  capabilities array.

- Remove an unused RsslError object in rsslVATest.

- In TunnelStream function createBigBuffer, cast the malloc parameter to
  a UInt32, matching the length set on the buffer.

- In JNI code that notifies a server socket, use InetPton instead of
  inet_addr on Windows. Windows has deprecated inet_addr.

- Remove a test that performs some comparisons of varying floats and
  doubles. It does not test any API functionality, and generates a
  warning about overflow (C4756) on Windows Release builds.

Fix rsslCopyMsg bugs:

- Fix a crash when calling rsslReleaseCopiedMsg on a non-AckMsg when
  RSSL_CMF_NAK_TEXT was among the flags used in rsslCopyMsg. Code
  attempted to free ackMsg.text.data.

- Fix rsslCopyMsg unsetting ReqMsgKey name/attrib flags based on whether
  the base MsgKey name/attrib are copied (should be based on the
  ReqMsgKey).

- Fix rsslReleaseCopiedMsg attempting to free the ReqMsgKey based on
  whether the base MsgKey was copied, triggering a memory error.

Add helpers for setting/checking the HAS_REQ_MSG_KEY flag for
Refresh/Status/Generic messages and use them in the tests.

Fix GenericMsg doxygen to say 'GenericMsg' instead of 'CloseMsg'.

Update testing of rsslCopyMsg in rsslMessageTest to include
API-allocated copies.  Fix instances where the tests check values of the
original message instead of the copy. Move buffers allocated for tests
such as these to the main function so that they are still initialized
when using --gtest_filter to run specific tests.  Add tests for
ReqMsgKey on Refresh, Status, and Generic messages.

Move deallocation of buffers in rsslMessageTest from a test to after
RUN_ALL_TESTS in main, so that they are cleaned up if gtest_filter  is
used.