ESDK-2551 Fixed java.nio.BufferOverflowException in encodeVectorEntry and encodeFilterEntry code.
In some edge cases the tracking of the bytes written would go out of synch and code would not correctly
detect the overflow condition. The correct behavior is to return the BUFFER_TOO_SMALL error code.
The fix corrects the missing updates to the counter of bytes written to correctly detect the overflow
condition and to return the BUFFER_TOO_SMALL error code instead of attempting to write into the buffer
and cause the BufferOverflowException to be thrown.