ESDK-1715 for ETAJ Provides ease of use with the session management feature to ETAC Reactor and Watchlist components for subscribing market data from Elektron RealTime in the Cloud. The feature handles token management by refreshing the access token from the EDP token service and send login reissue message to the connecting service provider with the token in order to keep the connection alive. In additional, the feature supports endpoint discovery by specifying a location (e.g. us-east or eu-west) to query for an endpoint (host and port) that provides auto failover from EDP RealTime Service Discovery.
The following is the list of interfaces for the feature

-       ReactorOptions provides options to override the default EDP token service
        and RealTime Service Discovery URLs.
-       ConsumerRole provide a member to specify client ID for an application
        making a request to the EDP token service.
-       ReactorAuthTokenEvent and ReactorAuthTokenInfo provide token information events
        via the ReactorAuthTokenEventCallback.
-       ReactorConnectInfo provides additional members as follows:
        The enableSessionManagement member is used to indicate RsslReactor to request the
        authentication token and query an endpoint from EDP-RT service discovery if both host
        and port are not specified by users.
        The location member is used to specifies a location to get a service endpoint to establish
        a connection with service provider. Defaults to "us-east" if not specified.
        The reactorAuthTokenEventCallback method is used to provide the ReactorAuthTokenEventCallback
        function to receive ReactorAuthTokenEvent. The access token is provided by the Reactor
        for Consumer(disabling watchlist) and NiProvider applications to send login request and
        reissue with the token.
-       The queryServiceDiscovery() is a synchronous method to query endpoints from EDP
        RealTime Service Discovery using the ReactorServiceDiscoveryOptions parameter and receive
        response via the ReactorServiceEndpointEventCallback callback method.

The VAConsumer and WatchlistConsumer provide additional parameters to support the session management
feature.

ESDK-1713 for EMAJ
Support the session management feature from the Reactor in EMA via programmatic methods, the EMA
configuration file and EMA programmatic configuration database. The user’s credentials are specified
by methods of the OmmConsumerConfig class while the Location and EnableSessionManagment parameters
are specified in the Channel group of EMA’s configuration.

The ServiceEndpointDiscovery is a standalone helper class to query endpoint information from EDP
RealTime Service Discovery using the ServiceEndpointDiscoveryOption and receive response via callback
methods of ServiceEndpointDiscoveryClient.

EMA provides the following consumer examples to demonstrate usage of the feature.

113__MarketPrice__SessionManagement illustrates how to use the EMA's configuration file to enable
session management and specify a location to get an endpoint for establishing a connection with
Elektron RealTime in the Cloud in cloud and consuming data.

450__MarketPrice__QueryServiceDiscovery illustrates how to query endpoints from EDP-RT service
discovery using the ServiceEndpointDiscovery class and use the location from the command line to
select an endpoint. The EMA's programmatic configuration is used to to enable session management
with the retrieved endpoint for establishing a connection with Elektron RealTime in the Cloud
and consuming data.