Avoid incorrect pattern of use of ReentrantLock.tryLock()
The broken pattern is

 try
 {
     if (lock.trylock())
     {
 ...
     }
 }
 catch (...)
 {
 ...
 }
 finally
 {
     lock.unlock();
 }

An equivalent correct pattern might be

 try
 {
     if (lock.trylock())
     {
 	try
         {
 ...
         }
         finally
         {
             lock.unlock();
         }
     }
 }
 catch (...)
 {
 ...
 }

For simplicity this commit does not preserve the exact same exception
handling. The code handling failure to lock used to be run within the
try-catch block, and isn't any more. However, since the try-catch blocks
seem to be there only to protect the case of successfully locking, this
shouldn't make any difference to the behaviour.