Squash merge of the following:
ESDK-3660: Added Websocket and JSON protocol support to the Consumer, Provider and NIProvider example applications.  These applications show how to use the ETA API to connect with a websocket connection, and to convert RWF to JSON and JSON to RWF.

ESDK-3654: Added Websocket and JSON protocol support to the ConsPer and ProvPerf examples.  These can now directly connect using websocket transport connections and support the JSON OMM Protocol.  In addition, enhanced the performance of the Date, Time, and DateTime to string functionality in ETA.

ESDK-3774 Fixed ETA to parse input buffer when there is one byte left in the input buffer. Added error handling and resizing of the reassemblyBuffer of rwsSession as needed to ensure that there is sufficient buffer to read data from websocket connection.