RTSDK-6136: Add in the multi-credential feature to ETAC and EMAC++
This feature allows users to associate specific login messages and oauth credentials with specific connections for Warm Standby and Channel List/Channel Set connections.

In addition, a new callback has been added to allow for changed userName and/or authenticationExtended login message elements when the Reactor or EMA recovers connections.

For the ETA Value Added Reactor, RsslRDMConsumerRole structures have an array of login messages and OAuth Credential structures.  The conneciton list connections and warm standby connections can configure login messages and OAuth credentials from the arrays in the RDMConsumerRole structure.

For EMA, OAuth credentials and Login Request messages can be configured on the OMMConsumerConfig class, with each credential and Login Request specifying a list of connections in the configuration to be used with those credentials.

For further information, please see the ETAC Value Added API Developers Guide, the EMA Config Guide, and the new EMAC++ Consumer Example 480 and ETAC's MultiCredWLConsumer.