RTSDK-6044: RTSDK Java Warm Standby Feature added. This adds the Warm Standby feature to both ETA Java and EMA Java.
RTSDK Warm Standby allows ETAJ Reactor Watchlist and EMAJ consuming applications to connect to multiple providers
with the same item watchlist to provide additional redundancy and recovery.  When configured, the API will connect to
multiple proivders, fan out requests on the behalf of the user to all of them, and provide responses and updates to the
application from the current active provider.  The API will communicating the active or standby status to the providers,
allowing for bandwidth savings as all responses from the standby providers will not be sent on the wire.  When a provider
disconnects or loses a service(if configured for Service-based Warm Standby), the API will transition to the proper
standby connection and alert the provider to resume sending data.

RTSDK Warm Standby supports either Login-based or Service-based Warm Standby.  Login-based Warm Standby is based on
connections, and the current active server will switch when the connection goes down.  Service-based Warm Standby will
manage the current active provider on a per-connection basis, switching the provider that is providing the active
service if a previous active service or connection goes down.

For more information, please see the ETAJ and EMAJ Developers guides, the EMAJ Configuration Guide.

For examples, please look at the Value Add wsbConsumer application for ETAJ and Cons470 for EMAJ.